# Cache generator types in registrar.